+++ rpc/config.go
@@ -1,6 +1,14 @@
 package rpc
 
-import pb "github.com/anycable/anycable-go/protos"
+import (
+	"crypto/tls"
+	"crypto/x509"
+	"errors"
+	"fmt"
+	"os"
+
+	pb "github.com/anycable/anycable-go/protos"
+)
 
 const (
 	defaultRPCHost = "localhost:50051"
@@ -27,6 +35,10 @@ type Config struct {
 	Concurrency int
 	// Enable client-side TLS on RPC connections?
 	EnableTLS bool
+	// Whether to verify the RPC server's certificate chain and host name
+	TLSVerify bool
+	// CA root TLS certificate path
+	TLSRootCA string
 	// Max receive msg size (bytes)
 	MaxRecvSize int
 	// Max send msg size (bytes)
@@ -43,5 +55,40 @@ type Config struct {
 
 // NewConfig builds a new config
 func NewConfig() Config {
-	return Config{Concurrency: 28, EnableTLS: false, Host: defaultRPCHost, Implementation: "grpc", RequestTimeout: 3000}
+	return Config{Concurrency: 28, EnableTLS: false, TLSVerify: true, Host: defaultRPCHost, Implementation: "grpc", RequestTimeout: 3000}
+}
+
+func (c *Config) TLSConfig() (*tls.Config, error) {
+	if !c.EnableTLS {
+		return nil, nil
+	}
+
+	var certPool *x509.CertPool = nil // use system CA certificates
+	if c.TLSRootCA != "" {
+		var rootCertificate []byte
+		var error error
+		if info, err := os.Stat(c.TLSRootCA); !os.IsNotExist(err) && !info.IsDir() {
+			rootCertificate, error = os.ReadFile(c.TLSRootCA)
+			if error != nil {
+				return nil, fmt.Errorf("failed to read RPC root CA certificate: %s", error)
+			}
+		} else {
+			rootCertificate = []byte(c.TLSRootCA)
+		}
+
+		certPool = x509.NewCertPool()
+		ok := certPool.AppendCertsFromPEM(rootCertificate)
+		if !ok {
+			return nil, errors.New("failed to parse RPC root CA certificate")
+		}
+	}
+
+	// #nosec G402: InsecureSkipVerify explicitly allowed to be set to true for development/testing
+	tlsConfig := &tls.Config{
+		InsecureSkipVerify: !c.TLSVerify,
+		MinVersion:         tls.VersionTLS12,
+		RootCAs:            certPool,
+	}
+
+	return tlsConfig, nil
 }